Conflicts:
	SoundStream/src/com/lastcrusade/soundstream/components/ConnectFragment.java

+++ SoundStream/src/com/lastcrusade/soundstream/components/PlaylistFragment.java
@@ -32,7 +32,7 @@ import android.view.MotionEvent;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup;
-import android.widget.ImageButton;
+import android.view.animation.TranslateAnimation;
 
 import com.lastcrusade.soundstream.CustomApp;
 import com.lastcrusade.soundstream.R;
@@ -41,7 +41,9 @@ import com.lastcrusade.soundstream.model.SongMetadata;
 import com.lastcrusade.soundstream.model.UserList;
 import com.lastcrusade.soundstream.service.PlaylistService;
 import com.lastcrusade.soundstream.service.ServiceLocator;
+import com.lastcrusade.soundstream.service.ServiceLocator.IOnBindListener;
 import com.lastcrusade.soundstream.service.ServiceNotBoundException;
+import com.lastcrusade.soundstream.service.UserListService;
 import com.lastcrusade.soundstream.util.BroadcastRegistrar;
 import com.lastcrusade.soundstream.util.IBroadcastActionHandler;
 import com.lastcrusade.soundstream.util.MusicListAdapter;
@@ -54,13 +56,19 @@ public class PlaylistFragment extends MusicListFragment{
     private BroadcastRegistrar registrar;
 
     private ServiceLocator<PlaylistService> playlistServiceServiceLocator;
+    private ServiceLocator<UserListService> userListServiceLocator;
 
     private PlayListAdapter mPlayListAdapter;
 
     @Override
     public void onCreate(Bundle savedInstanceState){
         super.onCreate(savedInstanceState);
-        final CustomApp curApp = (CustomApp) this.getActivity().getApplication();
+
+        mPlayListAdapter = new PlayListAdapter(
+                this.getActivity(),
+                Collections.<PlaylistEntry> emptyList(),
+                new UserList()
+        );
 
         playlistServiceServiceLocator = new ServiceLocator<PlaylistService>(
                 this.getActivity(),
@@ -75,12 +83,14 @@ public class PlaylistFragment extends MusicListFragment{
             }
         });
 
-        mPlayListAdapter = new PlayListAdapter(
-                this.getActivity(),
-                Collections.EMPTY_LIST,
-                curApp.getUserList()
-        );
-        setListAdapter(mPlayListAdapter);
+        userListServiceLocator = new ServiceLocator<UserListService>(
+                this.getActivity(), UserListService.class, UserListService.UserListServiceBinder.class);
+        userListServiceLocator.setOnBindListener(new IOnBindListener() {
+            @Override
+            public void onServiceBound() {
+                mPlayListAdapter.updateUsers(getUserListFromService());
+            }
+        });
 
         registerReceivers();
     }
@@ -89,7 +99,7 @@ public class PlaylistFragment extends MusicListFragment{
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
         View v = inflater.inflate(R.layout.list, container, false);
-       
+        setListAdapter(mPlayListAdapter);
         return v;
     }
 
@@ -103,6 +113,7 @@ public class PlaylistFragment extends MusicListFragment{
     public void onDestroy() {
         super.onDestroy();
         playlistServiceServiceLocator.unbind();
+        userListServiceLocator.unbind();
         unregisterReceivers();
     }
 
@@ -116,6 +127,14 @@ public class PlaylistFragment extends MusicListFragment{
         this.registrar.addAction(PlaylistService.ACTION_PLAYLIST_UPDATED, new IBroadcastActionHandler() {
             @Override
             public void onReceiveAction(Context context, Intent intent) {
+                updatePlaylist();
+            }
+        }).addAction(PlaylistService.ACTION_PLAYING_AUDIO, new IBroadcastActionHandler() {
+            @Override
+            public void onReceiveAction(Context context, Intent intent) {
+                //when the playlist starts playing a song, we want to make sure that we are
+                // showing the correct song being played, so we tell the adapter to update
+                // the playlist to force a redraw of the views
                 updatePlaylist();
             }
         })
@@ -157,7 +176,16 @@ public class PlaylistFragment extends MusicListFragment{
     private void updatePlaylist() {
         mPlayListAdapter.updateMusic(getPlaylistService().getPlaylistEntries());
     }
-    
+
+    private UserList getUserListFromService(){
+        UserList activeUsers = new UserList();
+        try {
+            activeUsers = userListServiceLocator.getService().getUserList();
+        } catch (ServiceNotBoundException e) {
+            Log.w(TAG, "UserListService not bound");
+        }
+        return activeUsers;
+    }
 
     private class PlayListAdapter extends MusicListAdapter<PlaylistEntry> {
         public PlayListAdapter(
@@ -174,13 +202,26 @@ public class PlaylistFragment extends MusicListFragment{
             
             //This depends on played music being above unplayed music
             PlaylistEntry entry = super.getItem(position);
+            
             if (!entry.isLoaded()) {
                 //TODO: style the unloaded elements here
-                element.setBackgroundColor(getResources().getColor(R.color.loading));
-            } else if (entry.isPlayed()) {
+                element.setBackgroundColor(getResources().getColor(R.color.abs__bright_foreground_disabled_holo_light));
+                element.findViewById(R.id.progress).setVisibility(View.VISIBLE);
+                element.findViewById(R.id.now_playing).setVisibility(View.INVISIBLE);
+            } 
+            else if (entry.isPlayed()) {
                 element.setBackgroundColor(getResources().getColor(R.color.used));
-            } else {
+                element.findViewById(R.id.now_playing).setVisibility(View.INVISIBLE);
+            } 
+            else {
                 element.setBackgroundColor(getResources().getColor(com.actionbarsherlock.R.color.abs__background_holo_light));
+                element.findViewById(R.id.progress).setVisibility(View.INVISIBLE);
+                if(entry.equals(getPlaylistService().getCurrentEntry())){
+                    element.findViewById(R.id.now_playing).setVisibility(View.VISIBLE);
+                }
+                else{
+                    element.findViewById(R.id.now_playing).setVisibility(View.INVISIBLE);
+                }
             }
             
 
@@ -193,10 +234,6 @@ public class PlaylistFragment extends MusicListFragment{
                 }
 
             });
-
-            ImageButton delete = (ImageButton)element.findViewById(R.id.btn_remove_from_playlist);
-            delete.setOnClickListener(new DeleteSongListener(entry));
-            delete.setVisibility(View.VISIBLE);
 
             
             return element;
@@ -213,37 +250,51 @@ public class PlaylistFragment extends MusicListFragment{
             }
         }
         
-        private class DeleteSongListener implements OnClickListener{
-            private PlaylistEntry entry;
-            public DeleteSongListener(PlaylistEntry entry){
-                this.entry = entry;
-            }
-            @Override
-            public void onClick(View v) {
-                if(getPlaylistService().getCurrentEntry()!= null && getPlaylistService().getCurrentEntry().equals(entry)){
-                    getPlaylistService().skip();
-                }
-                getPlaylistService().removeSong(entry);
-               
-            }
-            
-        }
-        
-
       //detect gestures 
         private class PlaylistSongGestureListener extends SongGestureListener{
             private PlaylistEntry entry;
+            private View view;
+            private final int SWIPE_MIN_DISTANCE = 100;
             
             public PlaylistSongGestureListener(View view, PlaylistEntry entry){
                 super(view);
+                this.view = view;
                 this.entry = entry;
+                
             }
             //fling a song to the right to remove it
             @Override
             public boolean onFling(MotionEvent e1, MotionEvent e2, float velocityX,
                     float velocityY) {
-                // TODO Implement remove this way in another pull request
-                return super.onFling(e1, e2, velocityX, velocityY);
+                boolean swipe = false;
+                // Fling is what the gesture detector detects
+                // Swipe is our internal vocabulary for a 
+                // horizontal left to right fling
+                if(isSwipe(e1, e2, velocityX, velocityY)){
+                    if(getPlaylistService().getCurrentEntry()!= null && 
+                            getPlaylistService().getCurrentEntry().equals(entry)){
+                        getPlaylistService().skip();
+                    }
+                    getPlaylistService().removeSong(entry);
+                    
+                    animateDragging((int)e2.getX());
+               
+                    swipe=true;
+                }
+                    
+                return swipe;
+            }
+            
+            //allows the view to be moved horizontally
+            @Override
+            public boolean onScroll(MotionEvent e1, MotionEvent e2,
+                float distanceX, float distanceY) {
+                
+                float dx = e2.getX() - e1.getX();
+                animateDragging(dx);
+                
+                return super.onScroll(e1, e2, distanceX, distanceY);
+            
             }
             
             //bump the song to the top when double tapped
@@ -252,6 +303,38 @@ public class PlaylistFragment extends MusicListFragment{
                 getPlaylistService().bumpSong(entry);
                 return true;
             } 
+            
+            /**
+             * Animates the current view by moving it to the right by the given 
+             * amount
+             * 
+             * @param amount
+             */
+            private void animateDragging(float amount){
+                TranslateAnimation trans = new TranslateAnimation(amount, amount, 0,0);
+                trans.initialize(view.getWidth(), view.getHeight(), 
+                        ((View)view.getParent()).getWidth(), ((View)view.getParent()).getHeight());
+                view.startAnimation(trans);
+            }
+            
+            /**
+             * Checks to see if the fling motion described by these inputs matches
+             * our definition of a left to right swipe
+             * 
+             * @param e1
+             * @param e2
+             * @param velocityX
+             * @param velocityY
+             * @return
+             */
+            private boolean isSwipe(MotionEvent e1, MotionEvent e2, float velocityX, float velocityY){
+                float dx = e2.getX() - e1.getX();
+                
+                if(dx > SWIPE_MIN_DISTANCE && velocityX > velocityY){
+                    return true;
+                }
+                return false;
+            }
         }
     }
